Fixes: ft_putx, ft_put_cx, ft_printf(.c .h)
Aesthetic:
Change the name of printf.h to ft_printf.h (The tester told me to, and i liked it.);
Also fixed that in all of the functions so there wouldn't be any misconvenior with the names etc;

Functions:
Changed the functions "ft_putx.c" "ft_put_cx.c" and "ft_printf.c". There was a mistake made by me.
I didn't RTFM, so i used "long long int" instead of "unsigned int", as it was said in the man
to do, so the buffer and the range were HUGE, compared to what it was meant to be. So i changed it,
but the compiler complained that the return type and the parameter type couldn't be the same (I still
don't know why.). Said so, i changed it "unsigned long int" INSIDE the "ft_putx.c" and "ft_put_cx.c"
functions. Check it out.

Also:
I tested at both testers, trippouile's and paulo-santana's, and passed them both after this changes.
The norm was tested and it's fine!
I'll send it today! (28/11). Wish me luck!

+++ ft_printf.c
@@ -6,11 +6,11 @@
 /*   By: igama <igama@student.42.rio>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 17:42:50 by igama             #+#    #+#             */
-/*   Updated: 2023/11/27 13:21:04 by igama            ###   ########.fr       */
+/*   Updated: 2023/11/28 08:58:00 by igama            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "printf.h"
+#include "ft_printf.h"
 
 int	ft_format(char c, va_list arguments)
 {
@@ -27,9 +27,9 @@ int	ft_format(char c, va_list arguments)
 	else if (c == 'p')
 		return (ft_digits_base(va_arg(arguments, unsigned long int)));
 	else if (c == 'x')
-		return (ft_digits_x(va_arg(arguments, long long int)));
+		return (ft_digits_x(va_arg(arguments, unsigned int)));
 	else if (c == 'X')
-		return (ft_digits_cx(va_arg(arguments, long long int)));
+		return (ft_digits_cx(va_arg(arguments, unsigned int)));
 	return (0);
 }
